Restructure lib/cli relationship
This change moves more of the concerns into the library (it seems better to have the library
enforce the ogg format requirement than have all callers mandate it). It still leaves the
same public API surface in case the caller does want to change the behavior but load now
handles the main concern (only ogg format past a certain point).

Also add more details to the wiki for getting ffmpeg successfully installed with the necessary
codecs etc. It is probably we will need to add more in the future (the ones included are the
ones I needed to successfully convert an mp3 I was testing with).